Move IPC out of Common.h/.cpp
