+++ CMakeLists.txt
@@ -83,6 +83,7 @@ add_executable(tev
     include/tev/ImageButton.h src/ImageButton.cpp
     include/tev/ImageCanvas.h src/ImageCanvas.cpp
     include/tev/ImageViewer.h src/ImageViewer.cpp
+    include/tev/Ipc.h src/Ipc.cpp
     include/tev/SharedQueue.h src/SharedQueue.cpp
     include/tev/ThreadPool.h src/ThreadPool.cpp
     include/tev/UberShader.h src/UberShader.cpp
